Moving student DTOs on web project
